Update stg_products.sql
in the products table, there are more than one value in attribute: supplier id in a format of #;# (eg: 2;6). Instead of removing the whole rows, it is determined to use the first value as the supplier id for the product because in the order detail table, such products were sold  (ed. product-id 74, 20). In order to save as much as data, the first value from the attribute supplier id will be treated as the supplier id without further information